// api/src/schemas_generated/output_pb.d.ts
export class ChatMessage extends jspb.Message {
  getSenderName(): string;
  setSenderName(value: string): void;

  getContent(): string;
  setContent(value: string): void;

  getMessageType(): string;
  setMessageType(value: string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ChatMessage.AsObject;
  static toObject(includeInstance: boolean, msg: ChatMessage): ChatMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: ChatMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ChatMessage;
  static deserializeBinaryFromReader(message: ChatMessage, reader: jspb.BinaryReader): ChatMessage;
}

export namespace ChatMessage {
  export type AsObject = {
    senderName: string,
    content: string,
    messageType: string,
  }
}

export class ChatHistory extends jspb.Message {
  clearMessagesList(): void;
  getMessagesList(): Array<ChatMessage>;
  setMessagesList(value: Array<ChatMessage>): void;
  addMessages(value?: ChatMessage, index?: number): ChatMessage;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ChatHistory.AsObject;
  static toObject(includeInstance: boolean, msg: ChatHistory): ChatHistory.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: ChatHistory, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ChatHistory;
  static deserializeBinaryFromReader(message: ChatHistory, reader: jspb.BinaryReader): ChatHistory;
}

export namespace ChatHistory {
  export type AsObject = {
    messagesList: Array<ChatMessage.AsObject>,
  }
}
// ...
export class AreaUpdate extends jspb.Message {
  getAreaId(): string;
  setAreaId(value: string): void;

  getName(): string;
  setName(value: string): void;

  getDescription(): string;
  setDescription(value: string): void;

  clearExitsList(): void;
  getExitsList(): Array<Exit>;
  setExitsList(value: Array<Exit>): void;
  addExits(value?: Exit, index?: number): Exit;

  clearItemsList(): void;
  getItemsList(): Array<AreaItem>;
  setItemsList(value: Array<AreaItem>): void;
  addItems(value?: AreaItem, index?: number): AreaItem;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): AreaUpdate.AsObject;
  static toObject(includeInstance: boolean, msg: AreaUpdate): AreaUpdate.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: AreaUpdate, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): AreaUpdate;
  static deserializeBinaryFromReader(message: AreaUpdate, reader: jspb.BinaryReader): AreaUpdate;
}

export namespace AreaUpdate {
  export type AsObject = {
    areaId: string,
    name: string,
    description: string,
    exitsList: Array<Exit.AsObject>,
    itemsList: Array<AreaItem.AsObject>,
  }
}
// ...
export class InventoryList extends jspb.Message {
  clearItemsList(): void;
  getItemsList(): Array<InventoryItem>;
  setItemsList(value: Array<InventoryItem>): void;
  addItems(value?: InventoryItem, index?: number): InventoryItem;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): InventoryList.AsObject;
  static toObject(includeInstance: boolean, msg: InventoryList): InventoryList.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: InventoryList, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): InventoryList;
  static deserializeBinaryFromReader(message: InventoryList, reader: jspb.BinaryReader): InventoryList;
}

export namespace InventoryList {
  export type AsObject = {
    itemsList: Array<InventoryItem.AsObject>,
  }
}

export class CharacterList extends jspb.Message {
  clearCharactersList(): void;
  getCharactersList(): Array<ListCharacter>;
  setCharactersList(value: Array<ListCharacter>): void;
  addCharacters(value?: ListCharacter, index?: number): ListCharacter;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): CharacterList.AsObject;
  static toObject(includeInstance: boolean, msg: CharacterList): CharacterList.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: CharacterList, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): CharacterList;
  static deserializeBinaryFromReader(message: CharacterList, reader: jspb.BinaryReader): CharacterList;
}

export namespace CharacterList {
  export type AsObject = {
    charactersList: Array<ListCharacter.AsObject>,
  }
}
// ...
export class CharacterSheet extends jspb.Message {
  getId(): string;
  setId(value: string): void;

  getName(): string;
  setName(value: string): void;

  getHealth(): number;
  setHealth(value: number): void;

  getMaxHealth(): number;
  setMaxHealth(value: number): void;

  getActionPoints(): number;
  setActionPoints(value: number): void;

  getMaxActionPoints(): number;
  setMaxActionPoints(value: number): void;

  hasCoreAttributes(): boolean;
  clearCoreAttributes(): void;
  getCoreAttributes(): CoreAttributes | undefined;
  setCoreAttributes(value?: CoreAttributes): void;

  hasDerivedStats(): boolean;
  clearDerivedStats(): void;
  getDerivedStats(): DerivedStats | undefined;
  setDerivedStats(value?: DerivedStats): void;

  getProficiencyLevel(): number;
  setProficiencyLevel(value: number): void;

  getPowerBudget(): number;
  setPowerBudget(value: number): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): CharacterSheet.AsObject;
  static toObject(includeInstance: boolean, msg: CharacterSheet): CharacterSheet.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: CharacterSheet, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): CharacterSheet;
  static deserializeBinaryFromReader(message: CharacterSheet, reader: jspb.BinaryReader): CharacterSheet;
}

export namespace CharacterSheet {
  export type AsObject = {
    id: string,
    name: string,
    health: number,
    maxHealth: number,
    actionPoints: number,
    maxActionPoints: number,
    coreAttributes?: CoreAttributes.AsObject,
    derivedStats?: DerivedStats.AsObject,
    proficiencyLevel: number,
    powerBudget: number,
  }
}
// ...
export namespace EquipmentUpdate {
  export type AsObject = {
    mainHand?: EquippedItem.AsObject,
    offHand?: EquippedItem.AsObject,
    head?: EquippedItem.AsObject,
    chest?: EquippedItem.AsObject,
    legs?: EquippedItem.AsObject,
    feet?: EquippedItem.AsObject,
  }
}

export class MetricsReport extends jspb.Message {
  getJsonPayload(): string;
  setJsonPayload(value: string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): MetricsReport.AsObject;
  static toObject(includeInstance: boolean, msg: MetricsReport): MetricsReport.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: MetricsReport, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): MetricsReport;
  static deserializeBinaryFromReader(message: MetricsReport, reader: jspb.BinaryReader): MetricsReport;
}

export namespace MetricsReport {
  export type AsObject = {
    jsonPayload: string,
  }
}
// ...
export namespace LoreCardCollection {
  export type AsObject = {
    cardsList: Array<LoreCardInstance.AsObject>,
  }
}
// ...
export class OutputEvent extends jspb.Message {
  clearTargetUserIdsList(): void;
  getTargetUserIdsList(): Array<string>;
  setTargetUserIdsList(value: Array<string>): void;
  addTargetUserIds(value: string, index?: number): string;

  hasChatHistory(): boolean;
  clearChatHistory(): void;
  getChatHistory(): ChatHistory | undefined;
  setChatHistory(value?: ChatHistory): void;

  hasChatMessage(): boolean;
  clearChatMessage(): void;
  getChatMessage(): ChatMessage | undefined;
  setChatMessage(value?: ChatMessage): void;

  hasAreaUpdate(): boolean;
  clearAreaUpdate(): void;
  getAreaUpdate(): AreaUpdate | undefined;
  setAreaUpdate(value?: AreaUpdate): void;

  hasCharacterList(): boolean;
  clearCharacterList(): void;
  getCharacterList(): CharacterList | undefined;
  setCharacterList(value?: CharacterList): void;

  hasCharacterSheet(): boolean;
  clearCharacterSheet(): void;
  getCharacterSheet(): CharacterSheet | undefined;
  setCharacterSheet(value?: CharacterSheet): void;

  hasInventoryList(): boolean;
  clearInventoryList(): void;
  getInventoryList(): InventoryList | undefined;
  setInventoryList(value?: InventoryList): void;

  hasMetricsReport(): boolean;
  clearMetricsReport(): void;
  getMetricsReport(): MetricsReport | undefined;
  setMetricsReport(value?: MetricsReport): void;

  hasEquipmentUpdate(): boolean;
  clearEquipmentUpdate(): void;
  getEquipmentUpdate(): EquipmentUpdate | undefined;
  setEquipmentUpdate(value?: EquipmentUpdate): void;

  hasLoreCardCollection(): boolean;
  clearLoreCardCollection(): void;
  getLoreCardCollection(): LoreCardCollection | undefined;
  setLoreCardCollection(value?: LoreCardCollection): void;

  hasLevelUpNotification(): boolean;
  clearLevelUpNotification(): void;
  getLevelUpNotification(): LevelUpNotification | undefined;
  setLevelUpNotification(value?: LevelUpNotification): void;

  hasLoreCardAwarded(): boolean;
  clearLoreCardAwarded(): void;
  getLoreCardAwarded(): LoreCardAwarded | undefined;
  setLoreCardAwarded(value?: LoreCardAwarded): void;

  getTraceId(): string;
  setTraceId(value: string): void;

  getPayloadCase(): OutputEvent.PayloadCase;
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): OutputEvent.AsObject;
  static toObject(includeInstance: boolean, msg: OutputEvent): OutputEvent.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: OutputEvent, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): OutputEvent;
  static deserializeBinaryFromReader(message: OutputEvent, reader: jspb.BinaryReader): OutputEvent;
}

export namespace OutputEvent {
  export type AsObject = {
    targetUserIdsList: Array<string>,
    chatHistory?: ChatHistory.AsObject,
    chatMessage?: ChatMessage.AsObject,
    areaUpdate?: AreaUpdate.AsObject,
    characterList?: CharacterList.AsObject,
    characterSheet?: CharacterSheet.AsObject,
    inventoryList?: InventoryList.AsObject,
    metricsReport?: MetricsReport.AsObject,
    equipmentUpdate?: EquipmentUpdate.AsObject,
    loreCardCollection?: LoreCardCollection.AsObject,
    levelUpNotification?: LevelUpNotification.AsObject,
    loreCardAwarded?: LoreCardAwarded.AsObject,
    traceId: string,
  }

  export enum PayloadCase {
    PAYLOAD_NOT_SET = 0,
    CHAT_HISTORY = 2,
    CHAT_MESSAGE = 3,
    AREA_UPDATE = 4,
    CHARACTER_LIST = 5,
    CHARACTER_SHEET = 6,
    INVENTORY_LIST = 7,
    METRICS_REPORT = 8,
    EQUIPMENT_UPDATE = 9,
    LORE_CARD_COLLECTION = 11,
    LEVEL_UP_NOTIFICATION = 12,
    LORE_CARD_AWARDED = 13,
  }
}